refactor moveTitle
